Promotion feature not tested

+++ cashierApp/src/main/java/alimentation/cashierApp/services/PromotionService.java
@@ -3,9 +3,12 @@ package alimentation.cashierApp.services;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import alimentation.cashierApp.dao.PromotionRepository;
+import alimentation.cashierApp.exception.CashierAppException;
+import alimentation.cashierApp.models.ProductType;
 import alimentation.cashierApp.models.Promotion;
 
 @Service
@@ -13,14 +16,16 @@ public class PromotionService {
 
     @Autowired
     private PromotionRepository PromotionRepository;
+    @Autowired
+    private ProductTypeService productTypeService;
 
 
     // Add methods for Promotion-related operations here
-    Iterable<Promotion> getAllPromotions(){
+    public Iterable<Promotion> getAllPromotions(){
         return PromotionRepository.findAll();
     }
 
-    Promotion getPromotionById(int id){
+    public Promotion getPromotionById(int id){
         Optional<Promotion> response = PromotionRepository.findById(id);
         if (response.isPresent()) {
             return response.get();
@@ -28,15 +33,25 @@ public class PromotionService {
         return null;
     }
 
-    void addPromotion(Promotion Promotion){
-        PromotionRepository.save(Promotion);
+    public Promotion getPromotionByProductName(String productName){ //get by type also possible
+        ProductType productType = productTypeService.getProductTypeByName(productName);
+        return PromotionRepository.findByProductType(productType);
     }
 
-    void updatePromotion(Promotion Promotion){
-        PromotionRepository.save(Promotion);
+    public Promotion addPromotion(Promotion Promotion){
+        return PromotionRepository.save(Promotion);
     }
 
-    void deletePromotion(int id){
+    public Promotion updatePromotion(Promotion Promotion){
+        return PromotionRepository.save(Promotion);
+    }
+
+    public Promotion deletePromotion(int id){
+        Optional<Promotion> target = PromotionRepository.findById(id);
+        if(target.get() == null){
+            throw new CashierAppException(HttpStatus.BAD_REQUEST, "Promotion missing");
+        }
         PromotionRepository.deleteById(id);
+        return target.get();
     }
 }
